+++ CHANGELOG.md
@@ -15,6 +15,7 @@ New features:
 * Split `Truffle::Interop.read` into `.read_member` and `.read_array_element`.
 * Drop `is_` prefix in `Truffle::Interop.is_array_element_*` predicates.
 * The old array strategy code has been removed and all remaining nodes converted to the new `ArrayStoreLibrary`.
+* `Truffle::Interop.hash_keys_as_members` has been added to treat a Ruby Hash as a polyglot object with the Hash keys as members.
 
 Bug fixes: